Uploaded progress on Weapons and Bullets
Made the shoot function into one - depending on the bullets per shot and
spread arc. An arc of 0 means a straight line of fire.

The weapon knows of its own bullet, instead of the player class keeping
track of this.

WeaponMaker is the start of an attempt to create weapon and bullet types
automagically and then assign them to prefabs.

Nothing is tested yet.